+++ app/src/main/java/de/baumann/browser/popups/Popup_bookmarks.java
@@ -332,9 +332,6 @@ public class Popup_bookmarks extends AppCompatActivity {
 
                             }
                         }).show();
-
-
-
                 return true;
             }
         });